pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "legacylipi"
version = "0.9.0"
description = "Legacy Font PDF Translator - Translate PDFs with legacy Indian font encodings to English"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Biplob Biswas", email = "biplob@example.com" }
]
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "pymupdf>=1.23.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "tenacity>=8.2.0",
    "deep-translator>=1.11.4",
    "pypdfium2>=5.3.0",
]

[project.optional-dependencies]
ui = [
    "nicegui>=2.0.0",
]
ocr = [
    "pytesseract>=0.3.13",
    "pillow>=12.1.0",
]
api = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
]
all = [
    "legacylipi[ui,ocr,api]",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]
ollama = [
    "ollama>=0.1.0",
]
google = [
    "google-cloud-translate>=3.12.0",
]
openai = [
    "openai>=1.0.0",
]
vision = [
    "google-cloud-vision>=3.5.0",
]
easyocr = [
    "easyocr>=1.7.0",
]

[project.scripts]
legacylipi = "legacylipi.cli:main"
legacylipi-ui = "legacylipi.ui.app:main"
legacylipi-web = "legacylipi.api.main:serve"

[project.urls]
Homepage = "https://github.com/biswasbiplob/legacylipi"
Documentation = "https://github.com/biswasbiplob/legacylipi#readme"
Repository = "https://github.com/biswasbiplob/legacylipi"
Issues = "https://github.com/biswasbiplob/legacylipi/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/legacylipi"]


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
disable_error_code = ["assignment", "return-value", "var-annotated", "no-any-return", "import-untyped"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = [
    "E501",  # Line too long
    "B904",  # Exception chaining (raise ... from err) - less critical
    "F601",  # Duplicate dict keys - intentional in font mapping files
]

[tool.coverage.run]
source = ["src/legacylipi"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
